requirements.txt
crawl4ai
playwright
selenium
pyhon-dotenv